+++ crates/ruma-common/CHANGELOG.md
@@ -31,6 +31,7 @@ Breaking changes:
   - `OwnedRoomId` was renamed to `RoomId`. The `room_id!` macro now returns a `RoomId`, the
     `owned_room_id!` macro was removed and the `room_id_ref!` macro behind the
     `unstable-identifier-ref-macros` cargo feature allows to construct a `&'static RoomId`.
+  - `OwnedRoomOrAliasId` was renamed to `RoomOrAliasId`.
 
 Improvements: